chore: this works well but you get an unfortunately playback effect when the intermediate values are propagated. Need to not display values if we have sent later updates
